Login,Register
Added registration and login using userName and password. I made a protectedRoute for the search and company name so that unauthorized users could not view stocks and add them to their portfolio. I logged out and displayed in the upper right corner the name of the user who visited the site